requirements.txt
alabaster==1.0.0
anyio==4.9.0
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.3.11
asttokens==3.0.0
async-lru==2.0.5
attractor_analysis==0.0.1
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.4
bleach==6.2.0
Brotli==1.1.0
cached-property==1.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
comm==0.2.2
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.1.0
fastjsonschema==2.21.1
fonttools==4.58.5
fqdn==1.5.1
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter==1.1.1
jupyter_client==8.6.3
jupyter-console==6.6.3
jupyter_core==5.7.2
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_server==2.16.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
line_profiler==4.2.0
llvmlite==0.44.0
lxml==6.0.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memory-profiler==0.61.0
mistune==3.1.3
myslideslive==0.7
myst-parser==4.0.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.4
notebook_shim==0.2.4
numba==0.61.2
numpy==2.2.6
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
pip==25.1.1
pkgutil_resolve_name==1.3.10
platformdirs==4.3.6
pluggy==1.6.0
prometheus_client==0.22.1
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.4.1
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.36.2
requests==2.32.4
reservoirpy==0.3.13.post1
rfc3339_validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.26.0
scikit-learn==1.7.0
scipy==1.15.3
Send2Trash==1.8.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.7
Sphinx==8.1.3
sphinx-autoapi==3.6.0
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tk==0.1.0
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20250708
typing_extensions==4.12.2
typing_utils==0.1.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
wheel==0.45.1
widgetsnbextension==4.0.14
zipp==3.23.0
zstandard==0.23.0
